requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==5.1.1
distro==1.9.0
Django==4.2
django-channels==0.7.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.0.1
Faker==28.4.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
HashTable==0.1
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.50.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
shortcuts==0.11.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
structures==0.9.5
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket==0.2.1
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.3
